Bug #11441
Fix the bug by overriding the WebRTC handler of the chat window init event.

Adds also a listener of events about room deletion so that the room is
automatically removed both from the user bookmarks and from his roster.

Now, when the owner of a room invte a user to join his room, the user becomes
a member of the room (in XMPP, this affiliation works only if it is done by the
owners or the admins of a room). With this information, it is now possible to
get the count of room members.

When a room is created, some properties are now explicitly set: the room is
hidden (not a public one) and it is not allowed for anyone to invite another
user.