Feature #11551
Fix some unit tests that use code handling datetime in i18n (and hence with
timezone) and in which the time offset wasn't correctly set. Unless Java 8,
Java 11 fixes by default the wrong offset according to the specified timezone.

Add Jaxb dependencies as it is removed in Java 11 (it was deprecated in
Java > 8).