Bug #11466
This fix provides a better solution to the previous urgent fix.

The bug is fixed by refactoring the way the SQL query is built. Instead of
fetching first the users matching some criteria, the user groups matching the
criterion on the group and on the roles are first fetched as those are very
much fewer than users. Then the SQL query is built by using inner selects and
joins. As the UserSearchCriteriaForDAO brings more complexities than it
resolves, it is now replaced by the SqlSelectorByCriteriaBuilder class whose
goal is to build the JdbcSqlQuery matching the criteria on the users. Same for
GroupSearchCriterionForDAO for the group search for better coherence with the
user search.

Fix also some minor hidden bugs and some sonarqube issues coming by new rules.