Feature #11551
Fix the code and the tests with the upgrade of some dependencies defined in the
top level definition of the Silverpeas projects (bom of dependencies).

For example, a bug has been fixed in iCal4J: when a recurrent event is
defined from a given date and for some days of week that doesn't match the
start date of the event, then now no occurrence of the event is occurring at
this start date as it doesn't match the recurrence rule.

Another example with Hibernate implementation of JPA: the getting of an entity
fails when that object requires (non-null dependency) another entity that has
been deleted, even with a lazy fetching strategy. Refine our own JPA base model
to be more compliant to Hibernate constraints on its own implementation of JPA.

Now Silverpeas is based upon JEE 8.

Specify the Docker container to use to execute the build with Java 11 and
Wildfly 20.

Add some unit tests.

In Jenkinsfile, compute the version of the parent pom when a snapshot version
of it is declared in the project's pom.