Feature #11551
Some integrations tests fail sometimes and randomly by a timeout in Arquillian
just because Wildfly takes more time with Java 11 to deploy the web archive of
the test. As a quick workaround, in the waiting of a better solution, the
Arquillian timeout is set at 2mn instead of the default 1mn.

In order to be JEE 8 compliant, the different XML descriptors were updated to
support this set of specification.

For better performance, we are targeting to configure the scan of managed beans
by CDI on only annotated classes. For doing, we have to annotate all our
classes that are subject to be managed by CDI. Therefore some of our
annotations were refined as being CDI stereotypes wrapping the life-cycle scope
of the annotated beans.

Remove in JUnit 5 tests the reference of the JUnit 4 API. Replaces in some
tests the deprecated JUnit 4 API functions by those provided by the  Hamcrest
API.