Feature #11354
Now, the chat service can be enabled for only some groups of users in
Silverpeas. By default, if no group identifiers are set in the chat.properties
settings, then the chat service is enabled for all the users in Silverpeas that
are in a domain mapped with an XMPP domain. Otherwises, only the users that are
in a mapped XMPP domain and in an allowed groups can use the chat service.