Feature #11354
Profit of the ACL capabilities of SilverChat to specify the groups of users in
Silverpeas that can create group chats and invite others users to join these
group chats.
For doing, a new property is added in the chat.properties settings file:
chat.acl.groupchat.creation. This property accepts as value a comma-separated
list of group identifiers. If not set, all users in Silverpeas can create group
chat and invite buddies to join them. If set, only the groups of users
specified in the property can have this functionality.