Feature #11841
NodeDetail implements now Identifiable. So, the getId() method is refined to
satisfy the Identifiable interface and the previous getId() method is renamed to
getLocalId().
Following this refining, the rightsDependOn property, that refers a node parent
from which the rights are derived, is now a String and not more an Integer (to
be compliant with the Identifiable interface).
BeanTranslation that represents a translation of a bean, and hence of a node
also, has now its identifier property represented by a String and not more by
an Integer; String has the interesting quality to represent any type of
identifiers (when the identifier isn't reified in the code) and this is why a
lot of beans in Silverpeas are encoded as a String. The technical detail of the
true type of the identifier of a bean has to be left to the DAOs.

NodeDetail implements now I18nContribution. In CMIS, we can use its
ContributionIdentifier as the identifier of its representative CMIS object.

The old Content Management API has been lightly refactored by renaming some of
its classes to a an unambiguous name according to their responsability and the
way they are used.

Now NodeDetail objects coming from Kmelia applications are exposed through our
CMIS objects tree implementation. The rights on the nodes are taken in charge.
Only normal nodes (therefore not the node of unclassed contributions and the
bin node) are exposed.

The user contributions are usually managed directly by and through the
component instances in Silverpeas. These applications use for doing the API
Silverpeas Core provides to handle transversal and generic business objects.
Our CMIS implementation has to know, at the level of the applications that are
exposed through the CMIS objects tree, what are the user contributions to also
expose in the tree and that are accessible to the current user. For doing, an
interface is provided by our CMIS implementation, CmisContributionsProvider,
for the components to implement if they want the contributions they manage to
be exposed through CMIS. Because the user contributions usually are built upon
the transversal or generic business objects, they are also automatically taken
in charge by the different CMIS tree walkers. These walkers known how to handle
and convert the business objets on which they work for their representation in
the CMIS objects tree.

Add a CMIS tree walker for publications and a new CMIS object type to represent
publications in Silverpeas.
For application using NodeDetails to categorize the user contributions, the
root node (root folder) is considered in our CMIS implementation as a folder
representation of the application; it doesn't have to be exposed and then when
the parent of a publication is the root folder, the Silverpeas's CMIS
implementation changes it as to be the application itself.