Feature #11841
In order to homogenize the getting of some of the common properties of the
resources managed in Silverpeas, some new interfaces are added:
* Nameable for all objects that have both a name and a short description;
* SilverpeasResource for all resources managed in Silverpeas and dedicated to
  the users;
* Translatable for all objects that supports the localization of their textual
  properties;
* LocalizedResource for all translatable and nameable resources in Silverpeas;
* ResourceTranslation for all translations of translatable and nameable
  resources in Silverpeas.
And the way the translations are handled in the AbstractI18NBean has been
reworked in a way the properties of such a bean can be get transparently both
in any language and in their default setting:
myBean.getTranslation(inALanguage).getName().
This way satisfies the new L10n API introduced with the Contribution API and
provides a bridge between this API with the old translation mechanism used in
Silverpeas.

The class BeanTranslation implements ResourceTranslation.
A base implementation of ResourceIdentifier is provided: BasicIdentifier.
SpaceInst(Light) and ComponentInst(Light) implement now the LocalizedResource
interface and use then a BasicIdentifier instance to represent their unique
identifier. Some of their methods have been renamed in order to be compliant
with the interface: getCreateDate to getCreationDate, getUpdateDate to
getLastUpdateDate and so on.

The interface Contribution has also be modified to take into account it extends
the SilverpeasResource interface (for example, the method getModifier has been
renamed to getLastUpdater).