Feature #11841
The attachments of a publication are now exposed as CMIS documents. The
publications act here as a CMIS folder of such objects. To handle more
easily the attachments in whatever language, a new class is introduced:
org.silverpeas.core.contribution.attachment.model.Document. This class is used
as a wrapper of all of the SimpleDocument instances that are a translation of
a single document (that can be represented by a SimpleDocument master).
The SimpleDocument class implements now the following interfaces:
LocalizedResource, ResourceTranslation and the new LocalizedAttachment that
extends itself the LocalizedContribution interface.

For instance, the documents attached to a publication are just exposed as such.
They cannot be read or modified.

The path computation has been improved by using the ResourcePath generic
interface: the computation is delegated to the different implementations of
this interface.

The CMIS representation of the resources and contributions in Silverpeas and
exposed in the CMIS objects tree extends now the ObjectDataImpl class of
Apache Chemistry.

SpaceInst, SpaceInstLight, and SilverpeasSharedComponentInstance implements now
the Securable interface.

Fix some bugs found with the Apache Chemistry Workbench tool.