Improve IT
RequestTaskManager is now a CDI bean. Improve the genericity declaration
in the classes of the Concurrent Task Managing API. All tasks are
cancelled at CDI is stopping.

The QuartzScheduler is now shutdown by CDI when unloaded (@PreDestroy
annotation)

Add database tables required by the integration tests in web-core as
well as the beans required by CDI (even their missing doesn't fail the
tests).

JCR schema is now installed by an Initializer and an initialization
listener is added for integration tests to bootstrap the Silverpeas JCR
schema.

Remove the test-ds.xml descriptors as the database is now defined in the
remote Wildfly