Adding a new XMPP Web Client : converse.js
Adding a silverpeas-converse.js plugin which provide the API to start the client.
The Silverpeas's functional rules are implemented.

The version 7.0.5 is used and provided by path /silverpeas/chat/converse.

A simple mechanism permits to choose between JSXC or ConverseJS client.

A new behavior has been addes to Silverpeas AJAX API. It is now possible to request an image as blob type. It is usefull to use it for XMPP VCard management with converse.js

Decoupling the SilverChat API from the Silverpeas User Session API by adding to the last one a machanism that permits to perform some treatments before a Silverpeas logout.

JSXC or ConverseJS sources are not parsed by the Silverpeas maven minifier plugin.