Feature #12036: adding minor/normal modification behavior.
Add integration test to check the chain between a publication update to
the saving of the event on the modification tracking works well.

Fix some issues detected by the above integration test. The transaction
in the ContributionTrackingEvent#save() is moved upto the tracking
service itself: no has to invoke directly and out a transaction such a
save() method; all has to be controlled by the tracking service itself.

Add unit test to check the tracking of the deletion of contributions.
Add integration test to check the persistence of the events on the
tracking of contribution modifications.

Add a new API for the tracking for changes in the contributions.
This API is made of a service to listen for changes in contributions, of
a repository to persist any events relative to those changes, a handler
of the modification context property about minor/major modification.

Add the sql scripts in busCore to add a new table to store the events
about the modifications of contributions.

The tracking for contribution modification API aims to register any
changes occurring on a contribution. But only concrete contributions
that are annotated with the @ModificationTracked are elected to be taken
in charge by this API. An Additional parameter, defined in the
org/silverpeas/contribution/settings/contribution.properties, and named
contribution.tracking.apps allows to specify for what applications the
tracking for changes is enabled. These two criteria have to be satisfy
for a contribution to be tracked for modifications.

With this new API, any modifications (creation, minor modifications,
major modifications, and deletions) are logged and then it allows to
construct an historic log of all changes occurring on a contribution
during its whole lifecycle. With such historic log, a manager could have
a clue of who and when has performed a change in a contribution between
two dates (for example between two major updates, or between a last
modification and the deletion).

Refactor the ContributionModificationContextHandler class to implement
the ContributionOperationContextPropertyHandler interface. This
interface is introduce to generalize the idea of handlers of a current
operation context implying a contribution. These handlers is to add
specific properties in the operation context in order to direct the
operation processor to different ways. For example,
ContributionModificationContextHandler is about the kind of modification
on a contribution: the property indicates then whether the modification
is a minor or a major one. According to the value of this property, the
update won't be implied the same mechanism (depends on the type of the
contribution and other stuff).