Feature #12036: adding minor/normal modification behavior.
Taking into account Nicolas's feedback about the UI.

Adding to transversal publication services the possibility to specify explicitly the kind of operation on an update (MOVE or UPDATE finally).

Inverting a condition in order to avoiding potential NullPointerException problems.

Fixing a problem around the publication event dispatched on an UPDATE (or a MOVE). The before data were not loaded from a new transaction which implies that the reloaded data wound be the one just updated by a caller in same transaction.
Adding some integration test in order to ensure that part.