Feature #12036: adding minor/normal modification behavior.
Add integration tests about the move.

Add support of contribution moves in the tracking for modifications of contribution.

Support the tracking for contributions creation.

Take care the case the modification of a contribution is done by a
batch process. In this case, when the modification is about an update,
there is no major nor minor updates but only a simple, classical update.
In this case, there is no explicit requester. So, either the last
updater property of the contribution is taken or the system user. For
doing, a system user has been introduced and can be get through the User
interface (and hence through the UserProvider implementation).

Go on the integration test on the process between the publication
service and the modification tracking service.