Feature #11841
Put the setting and the getting of CMIS properties into a single place:
CmisProperties so that the CMIS properties can be easily get when
creating a new object in the CMIS tree (and hence in Silverpeas).

Code the support of the Publication creation (as a CMIS folder) in our
CMIS implementation.
The adding of a new CMIS object into a CMIS folder is taken in charge by
the walker dedicated to the type of the CMIS folder and it delegates the
creation of the object to the walker in charge of the CMIS object to
create (and hence of the mapped Silverpeas object).
For example: the creation of a publication into a folder of a Kmelia
instance is done either by the  TreeWalkerForComponentInst instance
(for the root folder) or by the TreeWalkerForNodeDetail instance (for
any other inner folders) and these walkers delegate the creation of
the publication itself to the TreeWalkerForPublicationDetail walker.

Add a builder in SimpleAttachment to simplify the construction of such
instances.
Code the support of attaching a document to a publication through CMIS.

Fix an issue with the creation of a publication: some CMIS clients ask
for the creation of a folder instead of a publication as child of a
given folder or component instance although the creation of folders
isn't allowed through our CMIS support. Now, explicitly, when the
creation of a folder is asked, it is converted explicitly to a creation
of a publication (in the case the parent is a component instance or a
folder).

Add some guards in the creation of both folders and documents through
CMIS.

After adding a way to distinguish the different kinds of CMIS folders
of Silverpeas (space, application, ...) through the use of an UTF-8
symbol prefixing the object name, remove this mechanism symbols to
avoid some errors with the path of objects.

Documents (attachments) can now be updated directly. Take into account
a document can be locked for modification in Silverpeas. In this case,
the document can be opened through CMIS in read-only.