Bug #12443
Fix the bug that came from a double system notification about the update
of the attachment: one from the update of the file itself, the other
from the document unlocking. Such a notification is listening by the
publication update module that updates the last modification data of the
publication to which the document is attached.
The fix is to avoid such notification in the case of an update by
upload.

Add also a copy constructor to SimpleDocument and to SimpleAttachment
classes.

Disable quality analysis because sonarqube requires Java 11.