SSE - Improvements.
When receiving an SSE request, the first operation is to make the request asynchronous and then first events are sent.
Creating a manager dedicated to register/unregister and provide asynchronous request context.
Using ReentrantReadWriteLock in order to improve concurrency behavior.
Do not use thread pool to send the event by default.
Adding load test in order to verify SSE concurrency mechanism. Refactoring them in order to as less as possible the mock which are not accurate for load tests.