Bug #12479
Improve documentation on the PdC.

Add hashCode and equals methods to ClassifyPosition.

Now, the GlobalPdcManager#copyPositions method checks the classified
object has not yet the classiciation positions to copy.

Move the classe ApplicationService in Silverpeas Core API.

In PublicationTemplateManager#getPublicationTemplate(...): if the record
template isn't an IdentifiedRecordTemplate, then throw a well-informed
PublicationTemplateException instead of a leaving the ClassCastException
to be raised by the runtime.