Feature #12035: wiring the modification popins on massive publication update.
Making the transversal mechanism of popins on contribution modification working on a batch of resources instead of working only on a single one.
A new TAG has been implemented in order to manage a batch of contribution.

Adding the management of a location into mechanism which is checking existence of subscription. It is used for guessing rightly subscriptions into case of an alias.

Fixing a bad behavior on ImageFieldDisplayer which was playing the load of the HTML page as many times as it exists image field.

Cleaning rightly the transversal TAGs written in order to manage the build of a Silverpeas's HTML page.

The Javascript plugins can now working on a single resource or on a batch of resource.