Feature #12037: adding category management to favorite feature.
Adding to silverpeas.js sp.form.serialize, sp.form.serializeArray and sp.form.serializeJson tools.
Fixing also sp.selection.newCheckboxMonitor in order to handle the case where several batches of checkbox are displayed.

Adding new centrallized admin services for frontend in order to retrieve data about space and component instances.

Refactoring Request Router and Session Controller in order to use the new Web Controller Mechanism.

Adding the category management view and the UI to associate a link with a category.

Adding the possibility to set a category to several links at a same time.

Adding a VueJS component to handle the display into widget (like portlet or homepage for example).

Using the new VueJS plugin into portlet.

Adding new SQL scripts.

Adding the repository to manage the persistence of the categories associated with integration tests.

Adding the repository to manage the persistence of couples of link / category associated with integration tests.

Wiring the new repositories with the LinkDAO one and upgrading integration test.

Missing the user deletion behavior. Adding it with integration test.