Feature #11841
In order to easy the rollbacking to the previous state of the code (the
one without custom CMIS type symbol support), this change is in its own
commit.

After rethinking a way to handle correctly the path of the CMIS objects
in which each node is an eventual concatenation of the CMIS type symbol
and of the object l10n name, the symbols in each CMIS object type
dedicated to represent a Silverpeas resource type are reintroduced and
now the CMIS object name exposed in the CMIS tree is prefixed by such a
symbol.

The goal of such symbols is to represent what kind of CMIS type the
object is in the CMIS tree. Indeed, there is no way to attach an
icon to a custom CMIS type.