Feature #11841
Fix bugs:
- #12657
- #12656
- #12651
- #12653
- #12673
- #12677

Add new methods in SimpleDocument: isEdited() and isEditedBy(User).
SimpleDocument#isEdited() does what did SimpleDocument#isReadOnly(): it
checks the document is currently being edited by a user in Silverpeas.
So, now, SimpleDocument#isReadOnly() does what it is expected by its
name: the document is it read-only? It is read-only if the document is
either not edited by anyone but it cannot be modified by the current
user behind the scene or edited by another user than the current one.

Add a new method in CmisContributionsProvider:
getContribution(ContributionIdentifier, User) to get a contribution from
the service implementor. This method is now used both by
TreeWalkerForNodeDetail and TreeWalkerForPublicationDetail to get
respectively a given node and a given publication from their identifier.

TreeWalkerForSimpleDocument now uses WebdavService to open or to update
the content of a versionned document that is currently edited (id est
that is locked) by the current user behind the CMIS access. Any
versionned document that is not edited (then not locked) by the current
user is opened in read-only. If the Kmelia instance is versionned (in
this case the publications cannot be always visible), then a new created
document is versionned at a working version and it is automatically
locked to pursue its edition through CMIS.