Feature #12713
Add a basket to contain resources selected by a user. These resources
can be a contribution or an organizational resource. The far goal of the
basket is to replace the current clipboard used in copy/cut and paste
operations. For instance, it is dedicated to contain the publications a
user has selected to use for filling a given newsletter template.