Feature #12713
Add a REST-based web service to get, post and to pop items from the
selection basket. Improve some business code and the integration tests
on web resources accessible by a REST-based Web API. For example, a
method is added in RESTWebServiceTest to perform a user authentication
against Silverpeas in order to open an explicit session that can be used
in the integration test. For business code, the improve is about to use
contributions in a more generic way for the selection basket without
having to know their concrete type and to resolve some sonar feedbacks
about code quality.