Feature #12713
It is now possible to push incomplete resource data into the selection
basket through its Web RESTful API. The data are automatically reloaded
each time the content of the basket is returned by the web resource
representing the selection basket. Only the unique identifier and the
type of the resource is required.

Because the draftOutDate is always set at null and because the automatic
drafting out process (batch process) isn't more used (anyway it cannot
work because the draftOutDate is always null) since 2011, the process
and the draftOutDate attribute and its handling in the code are all
removed of the code.