Bug #12840: fixing restoring of a space which was not applying parent space profiles.
Fixing the spreading of space rights which was not merging correctly several space roles on a single instance role.

Fixing the non spreading of specific rights of a space to components and sub spaces when blocking its inheritance.

Adding a mechanism in charge of the lazy loading of some SpaceInst data:
* profiles
* sub spaces
* components
Wiring this mechanism into SpaceInst Object and into services in charge of the feeding of these data.

Fixing little sonarQube feedback.