Bug #12805
Add a new predefined template field valued with the URL of the server on
which Silverpeas is running: notification_base_serverurl

WebEncodeHelper is no more deprecated as it doesn't in a whole conflict
with Encode of OWASP. Indeed, it provides additional features for HTML
rendering of texts.

Convert NotificationParameterNames to an enum and update the code in
consequency.