Bug #12889: implementing @HEAD endpoints for ICal exports.
Adding also the calendar name into ICAL file content.

Fixing the export of events of own calendar from personal space which was not taking into account the events registered into main calendar.

Fixing calendar ICAL export about ATTENDEE in order to specify a MAILTO in any case (even if it does not exist an email for a user account).