Bug #12999: fixing the publication alias display when main location linked to removed component instance.
Implementing a checker able to indicate if a component instance or a space MUST be considered as removed.

Adding to ComponentAccessController the ability to determinate if a component instance is removed and using it into access control rules about component instances.

Using the above ability into PublicationAccessController and completing linked unit tests.

Refactoring PublicationDAO.getDistributionTree which had to take into account removed component instances.

Improving UI of Space and Component BIN behavior.