Bug #13019
Format of JID (XMPP identifiers) follows the RFC 7622.
According to this RFC, a JID is formatted as <localpart>@<domainpart>
for a user.
In Silverpeas, the localpart of the JID for a user is computed from his
login. So, when this login is an address email or contains a @
character, the localpart requires to be converted to be compliant with
the RFC. Previously, any part of a login that starts with such a special
character were removed from the login before being used as localpart of
a JID.
Now a new policy is implemented: any @ characters are replaced by a
specific code (for instance its UTF-8 hex equivalent but as word and not
as codepoint, id est 0x40). The policy of the JID formatting is defined
in the properties/org/silverpeas/chat/settings/chat.properties settings
file, under the property chat.xmpp.jid.policy. By default, the new
policy is used (value 1), as asked.