Bug #13050
Fix the bug. Now, when a user accesses a space or an application that has
been removed or for which his access rights have been removed, a
forbiddden error page is displayed with a button than send back the user
to the Silverpeas home page. Thisi error page is dedicated only to address
errors of accessing a space or an application in Silverpeas.