Add in MANIFEST.MF the dependency on resteasy-jaxb-provider to allow serialization/deserialization of XML. This is required for some of our users that wrote a custom Silverpeas component requesting some external web resources producing XML entities.
Add support of XML generation for the resource "policy of passwords" for
testing purpose.

Usually, the adding of this dependency isn't required if at least one
JAX-RS web service produces or consumes entities in XML. It is required
only when some clients access external web resources producing entities
in XML and there is not yet JAX-RS services working in XML in
Silverpeas. As we have added this capability to PasswordResource, this
dependency shouldn't be necessary. Nevertheless, I prefered to add it as
it is with the jackson-jaxrs-json-provider dependency.