Feature #12866
To allow Silverpeas projects to depend on core-api without having also to depend on its implementation
(core-lib), some features in core-lib are copied into core-api. For instance, the functions to get a User
by his login and domain, or by his API token are now declared in UserProvider.

Some codes in core-lib and in core-rs using Administration to find a User by some of his data
(login/domain, API token) are replaced by the use of UserProvider.

Update consequently the integration tests.