Feature #13197: implementing new community component.
After adding in the application component descriptor schema a new XML
element for specifying a mapping between the roles of an application
with those of the parent space, the user profile mapping mechanism for
component instances is now simplified by taking into account only the
mapping defined in the application component descriptors. The
roleMapping.properties file is not more used, thus it can be deleted.

Add a method in InheritableSpaceRoles to check a given role is supported
by collaborative spaces.

Improve EntityIdSetter for unit tests by using the Java Invoke API
instead of the Reflection one.

Fix the EnumNamedParameter so that is get the Enum value itself instead
of its String value.
Take into account the SonarQube feedback to refine the others parameter types

Improve the RESTWebService and RESTWebServiceTest

Add a SpaceHomePageType enum to define all the types of the supported
home page for a space.

In RESTWebServiceTest, when authorization is denied for all users in
Silverpeas, all the user profiles of all the component instances and
spaces defined in the test are removed.

Fix integration tests on JPA.

Take into account feedback of SonarQube