Feature #13179
Remove all the code relative to the deprecated view:comments tag.

Now the view tag displayComments uses the VueJS comments commponent to
render the block of comments on a given user contribution. The rewriting
of the comments widget by using VueJS allows us to separate the GUI
components from the logic behind.