Feature #13179
Instead of dispatching on Login.jsp, using directly /Login path when the aim is to redirect the user to the login page.

Renaming Login.jsp into welcome.jsp in order to avoid confusions.

Managing a new parameter with AutoRedirectServlet and Permalinks servlet that allow to force a user to login.

Taking into account alias case by identifying the contextual component instance against the one linked to the contribution.

Missing the use of v-html directive with commentText in order to display rightly the comment with wbr/> but no HTML.

Upgrading the security around comment update & delete actions.