Feature #13464
Take into account the feedback of sonar and IntellJ about the code
quality.

Add a new method to DataRecord to get the size of the data record in
term of fields.

Add a new method to Administrator to get all the profiles a user can
have in a given component instance (a profile here is a mapping between
a role and some privileges in a space or a component).