Feature #12866
The previous project Silverpeas JCR is now a subproject of Silverpeas Core.
The package org.silverpeas.jcr from Silverpeas JCR is renamed to org.silverpeas.core.jcr.

Below a synthesis of the work done on Silverpeas Core JCR.

- Add code to abstract the implementation of the JCR under the hoods.
In order to avoid to load any RepositoryFactory other than the ones provided
by the Silverpeas backend of the JCR, the interface SilverpeasRepositoryFactory is defined.
Any JCR implementation wrapper has to implement it.
For instance, Apache Jackrabbit Oak is used as the default implementation of the JCR.

- Add code to enable the configuration of the Oak repository from a
properties file. This is an alternative to the OSGi configuration in Oak
that has the drawback to depend on a lot of dependencies on both OSGi
and Felix (an implementation of OSGi).
The creation of the Oak repository is based upon the RepositoryFactory
that is loaded from the Java Service Provider as expected in the JCR
specifications. Silverpeas provides an implementation of this
RepositoryFactory interface. The creation of the Oak repository is
driven by the configuration loaded from a properties file whose the path
has to be provided by the system property jcr.conf. It is the goal of
the RepositoryFactory to create and configure the Oak repository from
the loaded configuration properties whereas it is the goal of the
RepositoryProvider to look for an available RepositoryFactory that knows
how to treat the peculiar Silverpeas parameters for the JCR. Any use of
the repository should ask for it to the RepositoryProvider which is a
singleton.

- The NodeStore in Oak requires to be closed/disposed at the ends of the
  application using the JCR in order to free some allocated resources (for
  example, the lock on the storage with the SegmentNodeStore.
  For doing, introduce the OakRepository and OakRepositoryConnection
  classes dedicated to wrap the Oak's repository implementation and to
  handle both the creation and the close of the repository (and hence of
  the NodeStore used behind). A new method is added to the interface
  SilverpeasRepositoryFactory: closeRepository; it is the symmetric
  operation of getting a repository in the RepositoryFactory interface.
  OakRepository extends SilverpeasRepository.

- Add a LoginModule to authenticate a Silverpeas user by login/password and
 another one by token.

- Add an initialization service to register a JCR schema for Silverpeas use.
Add unit tests on the JCR schema registering.

- Add a reentrant session mechanism:
For instance, a session can be opened within a single thread by only
one user, otherwise an exception is thrown. A session is distinguished
by the userID with which it is mapped. The reentrant mechanism is based
upon this distinction. For doing, based upon how the userID is fetched
by the session in the JCR implementations, a fallback mechanism is
implemented in the Silverpeas LoginModule for JCR to allow the userID
to be figuring out by the session.
In the Silverpeas wrapper of Oak, the userID is set at authentication
and identification success through an AuthInfo object.
The userID is the unique identifier of the user in Silverpeas.
The JCR system user is the virtual Silverpeas system user (id -1).

- Define an access controller for the JCR and this for whatever the
  implementation of the JCR used. Because, in the JCR, the access rights
  of the users and of the group of users, the authorization policies
  (ACL, ...) are defined in the content tree of the JCR, the
  implementations expect to find the data required to figuring out the
  privileges of the user as well as the permissions of the user on the
  accessed node directly from the JCR. But in Silverpeas access rights
  aren't stored in the JCR and to avoid a deduplication of the
  authorization mechanism between the JCR and Silverpeas, an access
  controller is defined with a simple rule of control, as it expects the
  control has been actually done by the AccessController objects during
  the navigation of the user in Silverpeas. So, each implementation of the
  JCR should use this simple access controller to perform their
  authorization checking.
  Implement for Oak a PermissionProvider and a TreePermission classes in which
  the control of access rights on the items in the JCR is performed as it was
  previously in Silverpeas by the project Silverpeas-JCR-AccessControl
  project.
  The access control is actually performed by the SilverpeasTreePermission
  instances. The SilverpeasPermissionProvider objects delegates the
  control to these instances. The implementation of the TreePermission
  extends the access controller presented above.

- Add WebDAV support to access documents in the JCR. This support was
  taken from Silverpeas-Core and adapted to this library. The WebDAV
  support is based upon the WebDAV JackRabbit implementation. A Link is
  done between an access token generated for the user accessing a document
  by WebDAV and his API token to ensure the authentication works fine.
  This is done transparently.
  The lock token generation in Oak is broken. It generates the token from
  the path of the accessed node in the JCR and as such the path contains
  problematic characters like the separator '/'.
  See https://issues.apache.org/jira/browse/OAK-10166
  A workaround of this issue is to wrap the incoming HTTP request in order
  to fix the token sent back by the WebDAV client so that the token value
  is the expected one by Oak.

- Add unit tests:
 * on the Oak repository configuration loading
 * on the authentications applied by the different
 * on the Oak repository login mechanism
 * on the creation of an Oak repository. For the tests, the repository is based upon a memory storage for the data
 * on the use of MongDB as storage for Oak.